move table to components
