+++ src/pages/contracts/index.tsx
@@ -3,7 +3,7 @@ import { ContractType } from '@/domain/repositories/DeploymentRepository'
 import { useListUserContracts } from '@/hooks/userContracts/useListUserContracts'
 import { EmptyString } from '@/services/common/EmptyString'
 import { FilterType } from '@/services/localDB/UserContractsRepository'
-import { ContractsTableContent } from '@/view/ContractView/ContractsTable'
+import { ContractsTableContent } from '@/view/components/ContractsTable'
 import MainContainer from '@/view/layout/MainContainer'
 import { Box, Paper, Typography } from '@mui/material'
 import { useState } from 'react'
