+++ src/view/components/ContractsTable/ContractsTable.tsx
