+++ src/view/components/ContractsTable/styled.ts
